latest

+++ src/scanner.c
@@ -14,15 +14,13 @@ static char        g_ch = '\0';
 static char        g_identifier[MAX_STRLEN] = {0};
 static int         g_number = 0;
 
-struct KeywordEntry
-{
+struct KeywordEntry {
     TokenKind kind;
     const char* identifier;
 };
 
 static struct KeywordEntry
-    g_keyword_table[] =
-{
+    g_keyword_table[] = {
     { TK_DIV,          "div"       },
     { TK_MOD,          "mod"       },
     { TK_LOGIC_OR,     "or"        },
@@ -50,10 +48,13 @@ inline static char
 get_char(void)
 {
     char c = *g_current;
+
     if(c == '\0')
         return c;
+
     if(c == '\n')
         g_line += 1;
+
     g_current += 1;
     return c;
 }
@@ -71,25 +72,29 @@ scan_identifier(void)
     assert(is_ident(g_ch));
     // identifier
     int i = 0;
+
     do {
-        if(i < MAX_STRLEN)
-        {
+        if(i < MAX_STRLEN) {
             g_identifier[i] = g_ch;
             i += 1;
         }
+
         g_ch = get_char();
     } while(is_ident(g_ch));
+
     g_identifier[i] = '\0';
     // keyword
     int k = 0;
+
     while(k < ARRAY_COUNT(g_keyword_table)
-           && !string_equal(g_keyword_table[k].identifier, g_identifier))
-    {
+          && !string_equal(g_keyword_table[k].identifier, g_identifier)) {
         k += 1;
     }
+
     if(k < ARRAY_COUNT(g_keyword_table)) {
         return g_keyword_table[k].kind;
     }
+
     return TK_IDENTIFIER;
 }
 
@@ -98,6 +103,7 @@ scan_decimal_number(void)
 {
     assert(is_digit(g_ch));
     long long value = 0;
+
     do {
         if(value > INT_MIN && value < INT_MAX) {
             value = 10 * value + g_ch - '0';
@@ -105,9 +111,11 @@ scan_decimal_number(void)
             scanner_mark_error("number too large for int");
             value = 0;
         }
+
         g_ch = get_char();
     } while(g_ch >= '0' && g_ch <= '9');
-    g_number = value;
+
+    g_number = (int)value;
     return TK_LITERAL_NUMBER;
 }
 
@@ -116,6 +124,7 @@ scan_hex_number(void)
 {
     assert(is_xdigit(g_ch));
     long long value = 0;
+
     do {
         if(value > INT_MIN && value < INT_MAX) {
             value = 16 * value + get_xdigit_value(g_ch);
@@ -123,9 +132,11 @@ scan_hex_number(void)
             scanner_mark_error("number too large for int");
             value = 0;
         }
+
         g_ch = get_char();
     } while(is_xdigit(g_ch));
-    g_number = value;
+
+    g_number = (int)value;
     return TK_LITERAL_NUMBER;
 }
 
@@ -133,20 +144,25 @@ static void
 skip_comment(void)
 {
     assert(g_ch == '*');
+
     do {
         // nested comments possible!
         do {
             g_ch = get_char();
+
             while(g_ch == '(') {
                 g_ch = get_char();
+
                 if(g_ch == '*')
                     skip_comment();
             }
         } while (g_ch != '*' || g_ch == '\0');
+
         do {
             g_ch = get_char();
         } while(g_ch == '*' || g_ch == '\0');
     } while(g_ch != ')' || g_ch == '\0');
+
     if(g_ch != '\0')
         g_ch = get_char();
     else
@@ -157,6 +173,7 @@ static void
 skip_line_comment(void)
 {
     assert(g_ch == '/');
+
     while(g_ch != '\n' && g_ch != '\0')
         g_ch = get_char();
 }
@@ -164,14 +181,15 @@ skip_line_comment(void)
 inline static void
 skip_whitespace(void)
 {
-    while(true){
-        switch(g_ch){
+    while(true) {
+        switch(g_ch) {
         case '\n':
         case ' ':
         case '\r':
         case '\t':
             g_ch = get_char();
             break;
+
         default:
             return;
         }
@@ -179,42 +197,159 @@ skip_whitespace(void)
 }
 
 
-TokenKind 
+TokenKind
 scanner_get(void)
 {
     skip_whitespace();
+
     if(g_ch == '\0')
         return TK_EOF;
+
     TokenKind kind = TK_UNKNOWN;
+
     if(is_ident_start(g_ch))
         kind = scan_identifier();
     else if(is_digit(g_ch))
         kind = scan_decimal_number();
     else {
         switch(g_ch) {
-        break;case '!': g_ch = get_char(); kind = scan_hex_number();
-        break;case '&': g_ch = get_char(); kind = TK_LOGIC_AND;
-        break;case '*': g_ch = get_char(); kind = TK_TIMES;
-        break;case '+': g_ch = get_char(); kind = TK_PLUS;
-        break;case '-': g_ch = get_char(); kind = TK_MINUS;
-        break;case '=': g_ch = get_char(); kind = TK_EQUAL;
-        break;case '#': g_ch = get_char(); kind = TK_NOT_EQUAL;
-        break;case ';': g_ch = get_char(); kind = TK_SEMICOLON;
-        break;case ',': g_ch = get_char(); kind = TK_COMMA;
-        break;case '.': g_ch = get_char(); kind = TK_PERIOD;
-        break;case '[': g_ch = get_char(); kind = TK_LEFT_BRACKET;
-        break;case ']': g_ch = get_char(); kind = TK_RIGHT_BRACKET;
-        break;case '~': g_ch = get_char(); kind = TK_LOGIC_NOT;
-        break;case '<': g_ch = get_char(); if(g_ch == '=') { kind = TK_LESS_EQUAL; g_ch = get_char(); } else { kind = TK_LESS; }
-        break;case '>': g_ch = get_char(); if(g_ch == '=') { kind = TK_GREATER_EQUAL; g_ch = get_char(); } else { kind = TK_GREATER; }
-        break;case ':': g_ch = get_char(); if(g_ch == '=') { kind = TK_ASSIGN; g_ch = get_char(); } else { kind = TK_COLON; }
-        break;case ')': g_ch = get_char(); kind = TK_RIGHT_PAREN;
-        break;case '(': g_ch = get_char(); if(g_ch == '*') { skip_comment(); kind = scanner_get(); } else { kind = TK_LEFT_PAREN; }
-        break;case '/': g_ch = get_char(); if(g_ch == '/') { skip_line_comment(); kind = scanner_get(); } else { unget_char(); }
-        break;default:
-        break;
+            break;
+
+        case '!':
+            g_ch = get_char();
+            kind = scan_hex_number();
+            break;
+
+        case '&':
+            g_ch = get_char();
+            kind = TK_LOGIC_AND;
+            break;
+
+        case '*':
+            g_ch = get_char();
+            kind = TK_TIMES;
+            break;
+
+        case '+':
+            g_ch = get_char();
+            kind = TK_PLUS;
+            break;
+
+        case '-':
+            g_ch = get_char();
+            kind = TK_MINUS;
+            break;
+
+        case '=':
+            g_ch = get_char();
+            kind = TK_EQUAL;
+            break;
+
+        case '#':
+            g_ch = get_char();
+            kind = TK_NOT_EQUAL;
+            break;
+
+        case ';':
+            g_ch = get_char();
+            kind = TK_SEMICOLON;
+            break;
+
+        case ',':
+            g_ch = get_char();
+            kind = TK_COMMA;
+            break;
+
+        case '.':
+            g_ch = get_char();
+            kind = TK_PERIOD;
+            break;
+
+        case '[':
+            g_ch = get_char();
+            kind = TK_LEFT_BRACKET;
+            break;
+
+        case ']':
+            g_ch = get_char();
+            kind = TK_RIGHT_BRACKET;
+            break;
+
+        case '~':
+            g_ch = get_char();
+            kind = TK_LOGIC_NOT;
+            break;
+
+        case '<':
+            g_ch = get_char();
+
+            if(g_ch == '=') {
+                kind = TK_LESS_EQUAL;
+                g_ch = get_char();
+            } else {
+                kind = TK_LESS;
+            }
+
+            break;
+
+        case '>':
+            g_ch = get_char();
+
+            if(g_ch == '=') {
+                kind = TK_GREATER_EQUAL;
+                g_ch = get_char();
+            } else {
+                kind = TK_GREATER;
+            }
+
+            break;
+
+        case ':':
+            g_ch = get_char();
+
+            if(g_ch == '=') {
+                kind = TK_ASSIGN;
+                g_ch = get_char();
+            } else {
+                kind = TK_COLON;
+            }
+
+            break;
+
+        case ')':
+            g_ch = get_char();
+            kind = TK_RIGHT_PAREN;
+            break;
+
+        case '(':
+            g_ch = get_char();
+
+            if(g_ch == '*') {
+                skip_comment();
+                kind = scanner_get();
+            } else {
+                kind = TK_LEFT_PAREN;
+            }
+
+            break;
+
+        case '/':
+            g_ch = get_char();
+
+            if(g_ch == '/') {
+                skip_line_comment();
+                kind = scanner_get();
+            } else {
+                unget_char();
+            }
+
+            break;
+
+        default:
+            break;
         }
     }
+
     return kind;
 }
 
@@ -243,7 +378,7 @@ scanner_get_number(void)
     return result;
 }
 
-void 
+void
 scanner_mark_error(const char* fmt, ...)
 {
     va_list arg;
